refactor(dataTableSkeleton): shift dataTableSkeleton under /Skeletons/common/DataTable
